// onyx-tasks-app/onyx/types.ts
// AUTO-GENERATED BY onyx-gen. DO NOT EDIT.
// Generated at 2026-01-13T08:13:59.423Z

export interface Task {
  id?: string;
  title?: string;
  description?: string;
  dueDate?: Date;
  completed?: boolean;
  [key: string]: any;
}

export type OnyxSchema = {
  Task: Task;
};

export type Schema = OnyxSchema;
export const Schema = {} as OnyxSchema;

export enum tables {
  Task = "Task",
}